+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## master
+
+* `EveOnline::ESI::CharacterAssetsLocations#etag` raise `NotImplementedError` from now
+* `EveOnline::ESI::CharacterAssetsNames#etag` raise `NotImplementedError` from now
+
 ## v0.24.0
 
 * Remove `implants` from `EveOnline::ESI::Models::JumpClone#as_json`