+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `EveOnline::ESI::CorporationAssetsLocations`
 * Add `EveOnline::ESI::CorporationAssetsNames`
 * Drop ruby 2.3 support
+* `EveOnline::ESI::Base#not_modified?` handle timeout exceptions
 
 ## v0.24.0